feat(cli): ✨ refactoring and detection of configs
