+++ packages/cli/src/types.d.ts
@@ -1,10 +1,14 @@
 
-interface Options {
-	environment: {
-		node: boolean
-		deno: boolean
-		browser: boolean
-	}
+export interface Config {
+	name: string
+	type: 'single' | 'multiple'
+	manual?: boolean
+	options: {
+		name: string
+		packages: Record<string, string | string[] | [string, string][]>
+		rules: string[]
+		detect?: string[] | true
+	}[]
 }
 
 export interface Workspace {
@@ -13,6 +17,8 @@ export interface Workspace {
 
 export interface Package {
 	name: string
+	path: string
 	files: string[]
 	directories: string[]
+	config?: Record<string, string[]>
 }
